Optimise switching of the video buffers
Previously, I called reset_drawing() to make sure the screen pointer
and line pointers were updated appropriately for the back buffer if
double buffering. I didn't realise that this carried a penalty of
calling memset on a 2.2 megabyte array to clear it. The intent was,
of course, to reset the emulated Amiga's video hardware entirely in
customreset.

This commit introduces a new function that simply sets the new screen
pointer and line pointers, and sdlgfx.cpp is made to call that.