Merge "Run" with the implied cancellation hotkey (B) in the main menu
Reset and Run are modified to read "Reset Amiga" and "Return to Amiga".
It is now clearer which option must be selected to reset the Amiga
and which option must be selected to simply return to the Amiga without
resetting it after loading a new disk, for example.

SaveStates and Filemanager are corrected.

Title Case options are modified to use regular case.